refactor(loop): 隔离步骤与 todo，重命名 loop_steps.step_id
本次变更包含两项紧密关联的重构：

### 1. step 处理器与 todo 完全隔离
将 list_steps / list_step_candidates / get_step / update_step / delete_step
从 handlers/todo.rs 移到独立的 handlers/step_.rs，todo.rs 只保留 promote。
路由也从 todo_routes() 分离到独立的 step_routes()。

### 2. loop_steps.todo_id → step_id 消除命名误导
loop_steps.todo_id 实际存储的是 steps.id（前端候选列表、后端创建验证、
引用计数均基于 steps 表），原名 todo_id 极具误导性。

变更涵盖：
- 实体层：field 名改为 step_id（DB 列名不变，通过 column_name 映射）
- 模型层：LoopStepRawDto / CreateLoopStepRequest / UpdateLoopStepRequest
  的字段均改为 step_id
- 数据库层：所有 Column::TodoId → Column::StepId
- 处理器层：req.todo_id / step.todo_id → req.step_id / step.step_id
- 前端层：类型定义和组件引用同步改为 step_id
