+++ backend/src/handlers/mod.rs
@@ -310,6 +310,7 @@ where
 
 mod todo;
 mod tag;
+mod step_;
 pub(crate) mod execution;
 mod scheduler;
 pub mod backup;
@@ -980,6 +981,7 @@ fn mount_domain_routes() -> Router<AppState> {
         .merge(cloud_routes())
         .merge(events_routes())
         .merge(loop_::loop_routes())
+        .merge(step_routes())
 }
 
 /// 给 TraceLayer 用的 span 工厂：把 `request_id` / `method` / `uri` 直接挂在 span 字段上，
@@ -1241,12 +1243,17 @@ fn todo_routes() -> Router<AppState> {
         // 环节相关：promote 复制到 steps 表,原 todo 保留
         .route("/api/todos/{id}/promote", post(todo::promote_todo_to_step))
         .route("/api/todos/{id}", get(todo::get_todo).put(todo::update_todo).delete(todo::delete_todo))
-        // 环节专用：list / candidates / 单查 / 更新,数据来自独立的 steps 表
-        .route("/api/steps", get(todo::list_steps))
-        .route("/api/steps/candidates", get(todo::list_step_candidates))
-        .route("/api/steps/{id}", get(todo::get_step).put(todo::update_step).delete(todo::delete_step))
+        // 环节专用路由已移至 step_routes()
         .route("/api/tags", get(tag::get_tags).post(tag::create_tag))
         .route("/api/tags/{id}", delete(tag::delete_tag))
+}
+
+/// 环节专用路由：数据来自独立的 steps 表，与 todo 完全隔离。
+fn step_routes() -> Router<AppState> {
+    Router::new()
+        .route("/api/steps", get(step_::list_steps))
+        .route("/api/steps/candidates", get(step_::list_step_candidates))
+        .route("/api/steps/{id}", get(step_::get_step).put(step_::update_step).delete(step_::delete_step))
 }
 
 /// 执行记录、执行触发、执行控制相关路由。
