+++ frontend/src/components/loop-flow/LoopFlowGraph.tsx
@@ -293,7 +293,7 @@ export function LoopFlowGraph({
                   fontSize={11}
                   fill="#64748b"
                 >
-                  {truncateText(node.step.todo_title || `#${node.step.todo_id}`, 22)}
+                  {truncateText(node.step.todo_title || `#${node.step.step_id}`, 22)}
                 </text>
                 <text
                   x={node.x + 12} y={node.y + 56}
